Introduce per-request UUIDs.
This commit adds a uuid field passed to all actions.
The UUID may be specified when firing the action or gets
generated as 'fetch-'{random}.

As part of this commit dispatchPayload is consistently merged
into all actions fired.

FetchConfiguration action was removed as it was replaced by
FetchInitiation with the last commit.